JAVA: Created a new package for Cloud and Server jira implementations. Added the entire descriptor for Jira Server so that it can be implemented in the future.
