+++ src/main/java/com/synopsys/integration/alert/channel/jira/server/descriptor/JiraServerDistributionUIConfig.java
@@ -1,0 +1,52 @@
+/**
+ * blackduck-alert
+ *
+ * Copyright (c) 2019 Synopsys, Inc.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package com.synopsys.integration.alert.channel.jira.server.descriptor;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import com.synopsys.integration.alert.channel.jira.server.JiraServerDescriptorKey;
+import com.synopsys.integration.alert.common.descriptor.config.field.CheckboxConfigField;
+import com.synopsys.integration.alert.common.descriptor.config.field.ConfigField;
+import com.synopsys.integration.alert.common.descriptor.config.ui.ChannelDistributionUIConfig;
+
+@Component
+public class JiraServerDistributionUIConfig extends ChannelDistributionUIConfig {
+    private static final String LABEL_ADD_COMMENTS = "Add comments";
+
+    private static final String DESCRIPTION_ADD_COMMENTS = "If true, this will add comments to the Jira ticket with data describing the latest change.";
+
+    @Autowired
+    public JiraServerDistributionUIConfig(final JiraServerDescriptorKey channelKey) {
+        super(channelKey, JiraServerDescriptor.JIRA_LABEL, JiraServerDescriptor.JIRA_URL);
+    }
+
+    @Override
+    public List<ConfigField> createChannelDistributionFields() {
+        ConfigField addCommentsField = CheckboxConfigField.create(JiraServerDescriptor.KEY_ADD_COMMENTS, LABEL_ADD_COMMENTS, DESCRIPTION_ADD_COMMENTS);
+
+        return List.of(addCommentsField);
+    }
+}
