+++ src/main/java/com/synopsys/integration/alert/channel/jira/server/descriptor/JiraServerGlobalUIConfig.java
@@ -1,0 +1,49 @@
+/**
+ * blackduck-alert
+ *
+ * Copyright (c) 2019 Synopsys, Inc.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package com.synopsys.integration.alert.channel.jira.server.descriptor;
+
+import java.util.List;
+
+import org.springframework.stereotype.Component;
+
+import com.synopsys.integration.alert.common.descriptor.config.field.ConfigField;
+import com.synopsys.integration.alert.common.descriptor.config.field.TextInputConfigField;
+import com.synopsys.integration.alert.common.descriptor.config.ui.UIConfig;
+
+@Component
+public class JiraServerGlobalUIConfig extends UIConfig {
+    private static final String LABEL_SERVER_URL = "Server URL";
+
+    private static final String DESCRIPTION_SERVER_URL = "The URL of the Jira server";
+
+    public JiraServerGlobalUIConfig() {
+        super(JiraServerDescriptor.JIRA_LABEL, JiraServerDescriptor.JIRA_DESCRIPTION, JiraServerDescriptor.JIRA_URL);
+    }
+
+    @Override
+    public List<ConfigField> createFields() {
+        ConfigField serverUrlField = TextInputConfigField.create(JiraServerDescriptor.KEY_SERVER_URL, LABEL_SERVER_URL, DESCRIPTION_SERVER_URL);
+
+        return List.of(serverUrlField);
+    }
+}
