+++ src/main/resources/db/changelog/alert/5.2.0/changelog-jira-server.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<databaseChangeLog xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+                   xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.3.xsd">
+
+    <changeSet id="1570824640421-1" author="bmandel">
+        <sql dbms="h2" stripComments="true">
+            CALL REGISTER_CHANNEL('channel_jira_server');
+        </sql>
+        <sql dbms="h2" stripComments="true">
+            CALL DEFINE_CHANNEL_FIELD('channel.jira.server.add.comments', 'channel_jira_server');
+
+            CALL DEFINE_FIELD('channel.jira.server.url', FALSE, 'channel_jira_server', 'GLOBAL');
+        </sql>
+    </changeSet>
+</databaseChangeLog>
